+++ spec/classes/init_spec.rb
@@ -10,18 +10,16 @@ describe 'snapd' do
       it { is_expected.to compile }
       it { is_expected.to contain_class('snapd') }
       it { is_expected.to contain_package('snapd') }
+
+      if fs_facts[:os]['name'] == 'Archlinux'
+        it {
+          is_expected.to contain_service('snapd.socket').with_ensure('running').with_enable(true).with_hasrestart(true).that_subscribes_to(['Package[snapd]'])
+        }
+      else
+        it {
+          is_expected.to contain_service('snapd').with_ensure('running').with_enable(true).with_hasrestart(true).that_subscribes_to(['Package[snapd]'])
+        }
+      end
     end
-  end
-
-  context 'on ubuntu' do
-    let(:facts) { {:os => { :family => 'Debian', :name => 'Ubuntu' } } }
-
-    it { is_expected.to contain_service('snapd').with_ensure('running').with_enable(true).with_hasrestart(true).that_subscribes_to(['Package[snapd]']) }
-  end
-
-  context 'on archlinux' do
-    let(:facts) { {:os => { :family => 'Archlinux', :name => 'Archlinux' } } }
-
-    it { is_expected.to contain_service('snapd.socket').with_ensure('running').with_enable(true).with_hasrestart(true).that_subscribes_to(['Package[snapd]']) }
   end
 end